requirements.txt
flask
python-dotenv
google-cloud-storage
google-cloud-firestore
google-cloud-vision